@@ -1,11 +1,10 @@
 name: moleculis
 description: Moleculis flutter application
 
-
 version: 0.3.0+10
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,6 +21,8 @@ dependencies:
   settings_ui: ^1.0.0
   image_picker: ^0.7.5
   cached_network_image: ^3.0.0
+  flexible_scrollbar: ^0.1.2
+  get_it: ^7.1.3
 
 dev_dependencies:
   flutter_test:
