@@ -1,12 +1,12 @@
 import 'package:moleculis/utils/format.dart';
 
 class CreateUpdateEventRequest {
-  final String title;
-  final String description;
-  final bool isPrivate;
-  final String location;
-  final DateTime date;
-  final List<String> users;
+  final String? title;
+  final String? description;
+  final bool? isPrivate;
+  final String? location;
+  final DateTime? date;
+  final List<String?>? users;
 
   CreateUpdateEventRequest({
     this.title,
@@ -23,18 +23,18 @@ class CreateUpdateEventRequest {
       'description': this.description,
       'isPrivate': this.isPrivate,
       'location': this.location,
-      'date': FormatUtils.formatDateTime(this.date),
+      'date': FormatUtils.formatDateTime(this.date!),
       'users': this.users,
     };
   }
 
   CreateUpdateEventRequest copyWith({
-    String title,
-    String description,
-    bool isPrivate,
-    String location,
-    DateTime date,
-    List<String> users,
+    String? title,
+    String? description,
+    bool? isPrivate,
+    String? location,
+    DateTime? date,
+    List<String?>? users,
   }) {
     return CreateUpdateEventRequest(
       title: title ?? this.title,
