@@ -1,11 +1,11 @@
 class LoginResponse {
-  final String token;
+  final String? token;
 
   LoginResponse({this.token});
 
   factory LoginResponse.fromMap(Map<String, dynamic> map) {
     return LoginResponse(
-      token: map['token'] as String,
+      token: map['token'] as String?,
     );
   }
 }
