@@ -1,29 +1,29 @@
-import 'package:moleculis/models/contact/contact.dart';
+import 'package:moleculis/models/contact.dart';
 import 'package:moleculis/models/enums/gender.dart';
-import 'package:moleculis/models/group/group.dart';
+import 'package:moleculis/models/group.dart';
 import 'package:moleculis/models/requests/update_user_request.dart';
 
 class User {
-  final String displayname;
-  final String fullname;
-  final String gender;
-  final List<Contact> contacts;
-  final List<Contact> contactRequests;
   final String username;
   final String email;
+  final String displayname;
+  final String fullname;
+  final Gender gender;
   final List<String> roles;
-  final List<Group> groups;
-  final List<Group> adminGroups;
+  final List<Contact>? contacts;
+  final List<Contact>? contactRequests;
+  final List<Group>? groups;
+  final List<Group>? adminGroups;
 
   User({
-    this.displayname,
-    this.fullname,
-    this.gender,
+    required this.username,
+    required this.email,
+    required this.displayname,
+    required this.fullname,
+    required this.gender,
+    required this.roles,
     this.contacts,
     this.contactRequests,
-    this.username,
-    this.email,
-    this.roles,
     this.groups,
     this.adminGroups,
   });
@@ -31,12 +31,12 @@ class User {
   factory User.fromMap(Map<String, dynamic> map) {
     final List<dynamic> senderContactsDynamic = map['contacts'] as List;
     final List<Contact> senderContacts =
-    senderContactsDynamic.map((i) => Contact.fromMap(i)).toList();
+        senderContactsDynamic.map((i) => Contact.fromMap(i)).toList();
 
     final List<dynamic> receiverContactsDynamic =
         map['contactRequests'] as List;
     final List<Contact> receiverContacts =
-    receiverContactsDynamic.map((i) => Contact.fromMap(i)).toList();
+        receiverContactsDynamic.map((i) => Contact.fromMap(i)).toList();
 
     final List<dynamic> rolesDynamic = map['roles'] as List;
 
@@ -51,7 +51,7 @@ class User {
     return User(
       displayname: map['displayname'] as String,
       fullname: map['fullname'] as String,
-      gender: map['gender'] as String,
+      gender: genderFromString(map['gender']),
       contacts: senderContacts,
       contactRequests: receiverContacts,
       username: map['username'] as String,
@@ -63,16 +63,16 @@ class User {
   }
 
   User copyWith({
-    String displayname,
-    String fullname,
-    String gender,
-    List<Contact> contacts,
-    List<Contact> contactRequests,
-    String username,
-    String email,
-    List<String> roles,
-    List<Group> groups,
-    List<Group> adminGroups,
+    String? displayname,
+    String? fullname,
+    Gender? gender,
+    List<Contact>? contacts,
+    List<Contact>? contactRequests,
+    String? username,
+    String? email,
+    List<String>? roles,
+    List<Group>? groups,
+    List<Group>? adminGroups,
   }) {
     return User(
       displayname: displayname ?? this.displayname,
@@ -90,11 +90,11 @@ class User {
 
   User copyWithRequest(UpdateUserRequest request) {
     return copyWith(
-        username: request.username,
-        email: request.email,
-        displayname: request.displayName,
-        fullname: request.fullName,
-        gender: request.gender.name
+      username: request.username,
+      email: request.email,
+      displayname: request.displayName,
+      fullname: request.fullName,
+      gender: request.gender,
     );
   }
 }
