@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:moleculis/blocs/authentication/authentication_bloc.dart';
-import 'package:moleculis/blocs/authentication/authentication_event.dart';
+import 'package:moleculis/blocs/auth/auth_bloc.dart';
+import 'package:moleculis/blocs/auth/auth_event.dart';
 import 'package:moleculis/models/user/user.dart';
 import 'package:moleculis/models/user/user_small.dart';
 import 'package:moleculis/screens/user_details/user_details_screen.dart';
@@ -12,14 +12,13 @@ class UserItem extends StatelessWidget {
   final User user;
 
   const UserItem({
-    Key key,
-    @required this.user,
+    Key? key,
+    required this.user,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final AuthenticationBloc authBloc =
-        BlocProvider.of<AuthenticationBloc>(context);
+    final AuthBloc authBloc = BlocProvider.of<AuthBloc>(context);
     return SimpleTile(
       onTap: () {
         Navigation.toScreen(
