@@ -1,23 +1,21 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:moleculis/models/user/user.dart';
-import 'package:moleculis/services/http_helper.dart';
-import 'package:moleculis/services/user_service.dart';
+import 'package:moleculis/services/apis/user_service.dart';
+import 'package:moleculis/utils/locator.dart';
 import 'package:moleculis/widgets/simple_tile.dart';
 
-class UserSearch extends SearchDelegate<User> {
+class UserSearch extends SearchDelegate<User?> {
   final List<String> excludeUsername;
 
-  UserSearch({@required this.excludeUsername});
+  UserSearch({required this.excludeUsername});
 
   @override
   List<Widget> buildActions(BuildContext context) {
     return [
       IconButton(
         icon: Icon(Icons.clear),
-        onPressed: () {
-          query = '';
-        },
+        onPressed: () => query = '',
       )
     ];
   }
@@ -25,10 +23,9 @@ class UserSearch extends SearchDelegate<User> {
   @override
   Widget buildLeading(BuildContext context) {
     return IconButton(
-        icon: Icon(Icons.arrow_back),
-        onPressed: () {
-          close(context, null);
-        });
+      icon: Icon(Icons.arrow_back),
+      onPressed: () => close(context, null),
+    );
   }
 
   @override
@@ -43,44 +40,43 @@ class UserSearch extends SearchDelegate<User> {
 
   Widget get _searchStream {
     return FutureBuilder<List<User>>(
-        future: UserService(HttpHelper()).getUsers(),
-        builder: (BuildContext context, AsyncSnapshot<List<User>> snapshot) {
-          if (!snapshot.hasData) {
-            return Center(
-              child: CircularProgressIndicator(),
-            );
-          }
-          final List<User> filteredResults = snapshot.data
-              .where((element) => !excludeUsername.contains(element.username))
-              .toList();
+      future: locator<UserService>().getUsers(),
+      builder: (BuildContext context, AsyncSnapshot<List<User>> snapshot) {
+        if (!snapshot.hasData) {
+          return Center(child: CircularProgressIndicator());
+        }
+        final List<User> filteredResults = snapshot.data!
+            .where((element) => !excludeUsername.contains(element.username))
+            .toList();
 
-          final List<User> searchResults = filteredResults.where((User user) {
-            return '${user.displayname} '
-                    '${user.fullname} '
-                    '${user.email} '
-                    '${user.username}'
-                .toLowerCase()
-                .contains(query.toLowerCase());
-          }).toList();
-          if (searchResults.isEmpty) {
-            return Center(
-              child: Text('no_res_found'.tr()),
-            );
-          }
-          return ListView.builder(
-            padding: const EdgeInsets.all(30),
-            itemCount: searchResults.length,
-            itemBuilder: (BuildContext context, int index) {
-              final user = searchResults[index];
-              return SimpleTile(
-                title: user.displayname,
-                subtitle: user.username,
-                onTap: () {
-                  close(context, user);
-                },
-              );
-            },
+        final List<User> searchResults = filteredResults.where((User user) {
+          return '${user.displayname} '
+                  '${user.fullname} '
+                  '${user.email} '
+                  '${user.username}'
+              .toLowerCase()
+              .contains(query.toLowerCase());
+        }).toList();
+        if (searchResults.isEmpty) {
+          return Center(
+            child: Text('no_res_found'.tr()),
           );
-        });
+        }
+        return ListView.builder(
+          padding: const EdgeInsets.all(30),
+          itemCount: searchResults.length,
+          itemBuilder: (BuildContext context, int index) {
+            final user = searchResults[index];
+            return SimpleTile(
+              title: user.displayname,
+              subtitle: user.username,
+              onTap: () {
+                close(context, user);
+              },
+            );
+          },
+        );
+      },
+    );
   }
 }
