@@ -20,10 +20,10 @@ class EventDetailsScreen extends StatefulWidget {
   final EventsBloc eventsBloc;
 
   const EventDetailsScreen({
-    Key key,
-    @required this.eventId,
-    this.owned,
-    this.eventsBloc,
+    Key? key,
+    required this.eventId,
+    required this.owned,
+    required this.eventsBloc,
   }) : super(key: key);
 
   @override
@@ -31,7 +31,7 @@ class EventDetailsScreen extends StatefulWidget {
 }
 
 class _EventDetailsScreenState extends State<EventDetailsScreen> {
-  EventsBloc eventsBloc;
+  late final EventsBloc eventsBloc;
 
   @override
   void initState() {
@@ -43,7 +43,7 @@ class _EventDetailsScreenState extends State<EventDetailsScreen> {
   Widget build(BuildContext context) {
     return BlocListener(
       bloc: widget.eventsBloc,
-      listener: (BuildContext context, state) {
+      listener: (BuildContext context, dynamic state) {
         if (state is EventsSuccess) {
           Navigator.pop(context);
         }
@@ -74,7 +74,7 @@ class _EventDetailsScreenState extends State<EventDetailsScreen> {
               child: BlocBuilder<EventsBloc, EventsState>(
                 bloc: eventsBloc,
                 builder: (BuildContext context, EventsState state) {
-                  final Event event = eventsBloc.getEventById(widget.eventId);
+                  final Event event = eventsBloc.getEventById(widget.eventId)!;
                   return Padding(
                     padding: const EdgeInsets.all(20),
                     child: Column(
@@ -85,7 +85,7 @@ class _EventDetailsScreenState extends State<EventDetailsScreen> {
                           child: BigTile(
                             title: event.title,
                             subtitle:
-                            '${'created'.tr()}: ${FormatUtils.formatDateAndTime(
+                                '${'created'.tr()}: ${FormatUtils.formatDateAndTime(
                               event.dateCreated,
                               context,
                             )}',
@@ -117,9 +117,7 @@ class _EventDetailsScreenState extends State<EventDetailsScreen> {
                         ),
                         InfoItem(
                           title: 'participants'.tr(),
-                          contentWidget: UsersList(
-                            users: event.users,
-                          ),
+                          contentWidget: UsersList(users: event.users),
                         ),
                       ],
                     ),
@@ -138,9 +136,10 @@ class _EventDetailsScreenState extends State<EventDetailsScreen> {
                   child: Text(
                     'leave_event'.tr(),
                     style: TextStyle(
-                        color: Colors.red,
-                        fontWeight: FontWeight.bold,
-                        fontSize: 20),
+                      color: Colors.red,
+                      fontWeight: FontWeight.bold,
+                      fontSize: 20,
+                    ),
                   ),
                 ),
               ),
