@@ -6,15 +6,13 @@ import 'package:moleculis/blocs/events/events_event.dart';
 import 'package:moleculis/blocs/events/events_state.dart';
 import 'package:moleculis/screens/create_edit_event/create_edit_event_screen.dart';
 import 'package:moleculis/screens/events/widgets/events_list.dart';
-import 'package:moleculis/services/events_service.dart';
-import 'package:moleculis/services/http_helper.dart';
 import 'package:moleculis/utils/navigation.dart';
 import 'package:moleculis/utils/widget_utils.dart';
 
 class EventsScreen extends StatefulWidget {
   final ShowSnackBar showErrorSnackBar;
 
-  const EventsScreen({Key key, @required this.showErrorSnackBar})
+  const EventsScreen({Key? key, required this.showErrorSnackBar})
       : super(key: key);
 
   @override
@@ -22,12 +20,11 @@ class EventsScreen extends StatefulWidget {
 }
 
 class _EventsScreenState extends State<EventsScreen> {
-  EventsBloc eventsBloc;
+  late final EventsBloc eventsBloc;
 
   @override
   void didChangeDependencies() {
-    eventsBloc = EventsBloc(eventsService: EventsService(HttpHelper()));
-    eventsBloc.add(LoadEvents());
+    eventsBloc = EventsBloc()..add(LoadEvents());
     super.didChangeDependencies();
   }
 
@@ -45,51 +42,49 @@ class _EventsScreenState extends State<EventsScreen> {
         child: DefaultTabController(
           length: 2,
           child: Scaffold(
-            appBar: WidgetUtils.appBar(context,
-                title: 'events'.tr().toLowerCase(),
-                bottom: TabBar(
-                  tabs: <Widget>[
-                    Tab(
-                      child: Text(
-                        'yours'.tr(),
-                        style: TextStyle(color: Colors.black),
-                      ),
+            appBar: WidgetUtils.appBar(
+              context,
+              title: 'events'.tr().toLowerCase(),
+              bottom: TabBar(
+                tabs: <Widget>[
+                  Tab(
+                    child: Text(
+                      'yours'.tr(),
+                      style: TextStyle(color: Colors.black),
                     ),
-                    Tab(
-                      child: Text(
-                        'others_events'.tr(),
-                        style: TextStyle(color: Colors.black),
-                      ),
-                    ),
-                  ],
-                ),
-                actions: [
-                  IconButton(
-                    icon: Icon(Icons.add),
-                    onPressed: () {
-                      Navigation.toScreen(
-                        context: context,
-                        screen: CreateEditEventScreen(
-                          eventsBloc: eventsBloc,
-                        ),
-                      );
-                    },
                   ),
-                ]),
+                  Tab(
+                    child: Text(
+                      'others_events'.tr(),
+                      style: TextStyle(color: Colors.black),
+                    ),
+                  ),
+                ],
+              ),
+              actions: [
+                IconButton(
+                  icon: Icon(Icons.add),
+                  onPressed: () {
+                    Navigation.toScreen(
+                      context: context,
+                      screen: CreateEditEventScreen(
+                        eventsBloc: eventsBloc,
+                      ),
+                    );
+                  },
+                ),
+              ],
+            ),
             body: BlocBuilder<EventsBloc, EventsState>(
               bloc: eventsBloc,
               builder: (BuildContext context, EventsState eventsState) {
                 if (eventsState.isLoading) {
-                  return Center(
-                    child: CircularProgressIndicator(),
-                  );
+                  return Center(child: CircularProgressIndicator());
                 }
                 return SafeArea(
                   child: TabBarView(
                     children: <Widget>[
-                      EventsList(
-                        events: eventsState.events,
-                      ),
+                      EventsList(events: eventsState.events),
                       EventsList(
                         events: eventsState.othersEvents,
                         others: true,
