@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:moleculis/blocs/groups/groups_bloc.dart';
-import 'package:moleculis/models/group/group.dart';
+import 'package:moleculis/models/group.dart';
 import 'package:moleculis/screens/group_details/group_details_screen.dart';
 import 'package:moleculis/utils/navigation.dart';
 import 'package:moleculis/widgets/simple_tile.dart';
@@ -9,7 +9,7 @@ import 'package:moleculis/widgets/simple_tile.dart';
 class GroupItem extends StatelessWidget {
   final Group group;
 
-  const GroupItem({Key key, @required this.group}) : super(key: key);
+  const GroupItem({Key? key, required this.group}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
