@@ -6,13 +6,13 @@ import 'package:flutter/material.dart';
 class ListRefresh extends StatelessWidget {
   final Widget child;
   final FutureOr onRefresh;
-  final bool isNoItems;
+  final bool? isNoItems;
   final String noItemsText;
 
   const ListRefresh({
-    Key key,
-    @required this.onRefresh,
-    @required this.child,
+    Key? key,
+    required this.onRefresh,
+    required this.child,
     this.isNoItems = false,
     this.noItemsText = '',
   }) : super(key: key);
@@ -20,8 +20,8 @@ class ListRefresh extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return RefreshIndicator(
-      onRefresh: onRefresh,
-      child: isNoItems
+      onRefresh: onRefresh as Future<void> Function(),
+      child: isNoItems!
           ? Center(
               child: Column(
                 mainAxisSize: MainAxisSize.min,
@@ -30,7 +30,7 @@ class ListRefresh extends StatelessWidget {
                   Padding(
                     padding: const EdgeInsets.only(top: 10),
                     child: GestureDetector(
-                      onTap: onRefresh,
+                      onTap: onRefresh as void Function()?,
                       child: Text(
                         'refresh'.tr(),
                         style: TextStyle(color: Colors.blue),
