@@ -3,19 +3,18 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:moleculis/blocs/groups/groups_event.dart';
 import 'package:moleculis/blocs/groups/groups_state.dart';
-import 'package:moleculis/models/group/group.dart';
+import 'package:moleculis/models/group.dart';
 import 'package:moleculis/models/page.dart';
 import 'package:moleculis/models/requests/create_update_group_request.dart';
 import 'package:moleculis/models/user/user_small.dart';
+import 'package:moleculis/services/apis/groups_service.dart';
 import 'package:moleculis/services/app_esceptions.dart';
-import 'package:moleculis/services/groups_service.dart';
+import 'package:moleculis/utils/locator.dart';
 
 class GroupsBloc extends Bloc<GroupsEvent, GroupsState> {
-  final GroupsService _groupsService;
+  final GroupsService _groupsService = locator<GroupsService>();
 
-  GroupsBloc({GroupsService groupsService})
-      : _groupsService = groupsService,
-        super(GroupsState());
+  GroupsBloc() : super(GroupsState());
 
   @override
   Stream<GroupsState> mapEventToState(GroupsEvent event) async* {
@@ -25,7 +24,11 @@ class GroupsBloc extends Bloc<GroupsEvent, GroupsState> {
       yield* _createGroup(event.request, event.users, event.admins);
     } else if (event is UpdateGroupEvent) {
       yield* _updateGroup(
-          event.request, event.users, event.admins, event.groupId);
+        event.request,
+        event.users ?? [],
+        event.admins ?? [],
+        event.groupId,
+      );
     }
   }
 
@@ -55,11 +58,11 @@ class GroupsBloc extends Bloc<GroupsEvent, GroupsState> {
   ) async* {
     try {
       yield state.copyWith(isLoading: true);
-      final List<String> userNames = users.map((e) => e.username).toList();
-      final List<String> adminNames = admins.map((e) => e.username).toList();
+      final List<String?> userNames = users.map((e) => e.username).toList();
+      final List<String?> adminNames = admins.map((e) => e.username).toList();
       request = request.copyWith(users: userNames, admins: adminNames);
 
-      final String message = await _groupsService.createGroup(request);
+      final String? message = await _groupsService.createGroup(request);
 
       yield GroupsSuccess(message: message);
       yield* _loadInitialData();
@@ -68,16 +71,21 @@ class GroupsBloc extends Bloc<GroupsEvent, GroupsState> {
     }
   }
 
-  Stream<GroupsState> _updateGroup(CreateUpdateGroupRequest request,
-      List<UserSmall> users, List<UserSmall> admins, int groupId) async* {
+  Stream<GroupsState> _updateGroup(
+    CreateUpdateGroupRequest request,
+    List<UserSmall> users,
+    List<UserSmall> admins,
+    int groupId,
+  ) async* {
     try {
       yield state.copyWith(isLoading: true);
-      final List<String> userNames = users.map((e) => e.username).toList();
-      final List<String> adminNames = admins.map((e) => e.username).toList();
+      final List<String?> userNames = users.map((e) => e.username).toList();
+      final List<String?> adminNames = admins.map((e) => e.username).toList();
       request = request.copyWith(users: userNames, admins: adminNames);
 
-      final String message = await _groupsService.updateGroup(request, groupId);
-      Group newGroup = getGroupById(groupId);
+      final String? message =
+          await _groupsService.updateGroup(groupId, request);
+      Group newGroup = getGroupById(groupId)!;
       newGroup = newGroup.copyWithRequest(request, users, admins);
       final List<Group> newGroups = state.groups;
       for (int i = 0; i < newGroups.length; ++i) {
@@ -97,7 +105,7 @@ class GroupsBloc extends Bloc<GroupsEvent, GroupsState> {
     }
   }
 
-  Group getGroupById(int groupId) {
+  Group? getGroupById(int groupId) {
     for (final group in state.groups) {
       if (group.id == groupId) {
         return group;
