@@ -2,19 +2,24 @@ import 'package:moleculis/models/user/user_small.dart';
 
 class Contact {
   final int id;
-  final bool accepted;
-  final bool isSender;
   final UserSmall user;
+  final bool isSender;
+  final bool accepted;
 
-  Contact({this.id, this.accepted, this.user, this.isSender});
+  Contact({
+    required this.id,
+    required this.user,
+    required this.isSender,
+    required this.accepted,
+  });
 
   factory Contact.fromMap(Map<String, dynamic> map) {
     final bool receiver = map['sender'] != null;
     return Contact(
       id: map['id'] as int,
-      accepted: map['accepted'] as bool,
-      isSender: !receiver,
       user: UserSmall.fromMap(receiver ? map['sender'] : map['receiver']),
+      isSender: !receiver,
+      accepted: map['accepted'] as bool,
     );
   }
 }
