@@ -1,9 +1,11 @@
 import 'package:moleculis/models/page.dart';
 import 'package:moleculis/models/requests/create_update_event_request.dart';
+import 'package:moleculis/services/apis/events_service.dart';
 import 'package:moleculis/services/http_helper.dart';
+import 'package:moleculis/utils/locator.dart';
 
-class EventsService {
-  final HttpHelper _httpHelper;
+class EventsServiceImpl implements EventsService {
+  final HttpHelper _httpHelper = locator<HttpHelper>();
   final String _endpointBase = '/events';
 
   String get _createEventEndpoint => _endpointBase + '/';
@@ -13,24 +15,27 @@ class EventsService {
   String _getOthersEventsPageEndpoint(int page) =>
       _endpointBase + '/others/page/$page';
 
-  String _updateEventEndpoint(int eventId) => _endpointBase + '/$eventId';
+  String _updateEventEndpoint(int? eventId) => _endpointBase + '/$eventId';
 
-  String _leaveEventEndpoint(int eventId) => _endpointBase + '/leave/$eventId';
+  String _leaveEventEndpoint(int? eventId) => _endpointBase + '/leave/$eventId';
 
-  EventsService(this._httpHelper);
-
+  @override
   Future<Page> getEventsPage(int page) async {
     final response = await _httpHelper.get(_getEventsPageEndpoint(page));
     return Page.fromMap(response);
   }
 
+  @override
   Future<Page> getOthersEventsPage(int page) async {
     final response = await _httpHelper.get(_getOthersEventsPageEndpoint(page));
     return Page.fromMap(response);
   }
 
-  Future<String> updateEvent(int eventId,
-      CreateUpdateEventRequest request) async {
+  @override
+  Future<String?> updateEvent(
+    int? eventId,
+    CreateUpdateEventRequest request,
+  ) async {
     final response = await _httpHelper.put(
       _updateEventEndpoint(eventId),
       body: request.toMap(),
@@ -38,7 +43,8 @@ class EventsService {
     return response['message'];
   }
 
-  Future<String> createEvent(CreateUpdateEventRequest request) async {
+  @override
+  Future<String?> createEvent(CreateUpdateEventRequest request) async {
     final response = await _httpHelper.post(
       _createEventEndpoint,
       body: request.toMap(),
@@ -46,7 +52,8 @@ class EventsService {
     return response['message'];
   }
 
-  Future<String> leaveEvent(int eventId) async {
+  @override
+  Future<String?> leaveEvent(int? eventId) async {
     final response = await _httpHelper.post(
       _leaveEventEndpoint(eventId),
     );
