@@ -1,15 +1,18 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:moleculis/common/img.dart';
+import 'package:moleculis/storage/shared_pref_manager.dart';
+import 'package:moleculis/utils/locator.dart';
 
 class LocaleUtils {
   static final String localePath = 'assets/locales';
+
+  static final defaultLocale = localeItems.first.locale;
+
   static final List<Locale> locales = [
     localeItems[0].locale,
     localeItems[1].locale,
   ];
-
-  static final List<String> localesLanguageCodes = ['en', 'uk'];
 
   static final List<LocaleItem> localeItems = [
     LocaleItem(
@@ -19,26 +22,40 @@ class LocaleUtils {
     ),
     LocaleItem(
       locale: Locale(LanguageCodes.uk.text),
-      name: 'Ukrainian',
+      name: 'Українська',
       imageAsset: Img.uaFlag,
     ),
   ];
 
   static LocaleItem currentLocaleItem(BuildContext context) {
     return localeItems.firstWhere((element) =>
-    element.locale.languageCode == context.locale.languageCode);
+        element.locale.languageCode == context.locale.languageCode);
   }
 
   static LocaleItem localeFlag(Locale locale) {
     return localeItems.firstWhere(
-            (element) => element.locale.languageCode == locale.languageCode);
+      (element) => element.locale.languageCode == locale.languageCode,
+    );
+  }
+
+  static Locale localeFromCode(String? langCode) {
+    switch (langCode) {
+      case 'en':
+        return localeItems.first.locale;
+      case 'uk':
+        return localeItems[1].locale;
+    }
+    return defaultLocale;
+  }
+
+  static Future<void> setLocale(BuildContext context, Locale locale) async {
+    if (!locales.contains(locale)) return;
+    locator<SharedPrefManager>().saveCurrentLocale(locale);
+    await context.setLocale(locale);
   }
 }
 
-enum LanguageCodes {
-  en,
-  uk,
-}
+enum LanguageCodes { en, uk }
 
 extension LanguageCodesExtension on LanguageCodes {
   String get text {
@@ -48,7 +65,6 @@ extension LanguageCodesExtension on LanguageCodes {
       case LanguageCodes.uk:
         return 'uk';
     }
-    return null;
   }
 }
 
@@ -58,8 +74,8 @@ class LocaleItem {
   final String imageAsset;
 
   LocaleItem({
-    this.locale,
-    this.name,
-    this.imageAsset,
+    required this.locale,
+    required this.name,
+    required this.imageAsset,
   });
 }
